Move directives to client dir
