+++ src/client/directives/TrimWhitespace.ts
